Move reusable helper methods to CommandAction
